Rewrote MenuDrawer and relocated navigation folder to src/ts/global
